ConstProp: Merge tests in loki.transformation.
