+++ loki/transformations/tests/test_constant_propagation.py
@@ -7,13 +7,98 @@
 
 import pytest
 
-from loki import FindNodes, Subroutine
-from loki.frontend import available_frontends
-from loki.ir import Assignment, Conditional, Loop
-from loki.jit_build import jit_compile
+from loki import Subroutine, available_frontends, jit_compile
+from loki.expression import symbols as sym
+from loki.ir import nodes as ir, FindNodes
 
+from loki.analyse import ConstantPropagationAnalysis, ConstantPropagationMapper
 from loki.transformations.constant_propagation import ConstantPropagationTransformer
 from loki.transformations.transform_loop import LoopUnrollTransformer
+
+
+def test_constant_propagation_analysis_declarations_map():
+    fcode = """
+subroutine const_prop_decls
+  integer :: a = 1
+  integer :: b(2) = (/2, 3/)
+  logical :: l = .true.
+end subroutine const_prop_decls
+    """.strip()
+    routine = Subroutine.from_source(fcode)
+
+    declarations_map = ConstantPropagationAnalysis().generate_declarations_map(routine)
+
+    assert declarations_map[('a', ())] == sym.IntLiteral(1)
+    assert declarations_map[('l', ())] == sym.LogicLiteral(True)
+    assert declarations_map[('b', (sym.IntLiteral(1),))] == sym.IntLiteral(2)
+    assert declarations_map[('b', (sym.IntLiteral(2),))] == sym.IntLiteral(3)
+
+
+def test_constant_propagation_mapper_folds_expressions():
+    mapper = ConstantPropagationMapper()
+
+    assert mapper(sym.Sum((sym.IntLiteral(1), sym.IntLiteral(2)))) == sym.IntLiteral(3)
+    assert mapper(sym.Quotient(sym.IntLiteral(7), sym.IntLiteral(2))) == sym.IntLiteral(3)
+    assert mapper(sym.Power(sym.IntLiteral(2), sym.IntLiteral(3))) == sym.IntLiteral(8)
+    assert mapper(sym.LogicalAnd((sym.LogicLiteral(True), sym.LogicLiteral(False)))) == sym.LogicLiteral(False)
+    assert mapper(sym.StringConcat((sym.StringLiteral('foo'), sym.StringLiteral('bar')))) == sym.StringLiteral('foobar')
+    assert mapper(sym.Sum((sym.FloatLiteral('1.5'), sym.FloatLiteral('2.5')))) == sym.FloatLiteral('4.0')
+
+
+def test_constant_propagation_mapper_short_circuits_boolean_ops():
+    mapper = ConstantPropagationMapper()
+    dyn = sym.Variable(name='dyn')
+
+    assert mapper(sym.LogicalOr((sym.LogicLiteral(True), dyn))) == sym.LogicLiteral(True)
+    assert mapper(sym.LogicalAnd((sym.LogicLiteral(False), dyn))) == sym.LogicLiteral(False)
+
+
+@pytest.mark.parametrize('frontend', available_frontends())
+def test_constant_propagation_analysis_attaches_maps(frontend):
+    fcode = """
+subroutine const_prop_attach
+  integer :: a = 1
+  integer :: b
+  b = a
+end subroutine const_prop_attach
+    """.strip()
+    routine = Subroutine.from_source(fcode, frontend=frontend)
+    assignments = FindNodes(ir.Assignment).visit(routine.body)
+
+    analysis = ConstantPropagationAnalysis()
+    analysis.attach_dataflow_analysis(routine)
+
+    assert assignments[0]._constants_map == {('a', ()): sym.IntLiteral(1)}
+
+    analysis.detach_dataflow_analysis(routine)
+    assert assignments[0]._constants_map is None
+
+
+def test_constant_propagation_analysis_dynamic_array_invalidation():
+    fcode = """
+subroutine const_prop_dynamic_array(a, i)
+  integer, intent(inout) :: a(3)
+  integer, intent(in) :: i
+  a(1) = 1
+  a(2) = 2
+  a(i) = 5
+end subroutine const_prop_dynamic_array
+    """.strip()
+    routine = Subroutine.from_source(fcode)
+    assignments = FindNodes(ir.Assignment).visit(routine.body)
+
+    analysis = ConstantPropagationAnalysis()
+    constants_map = {
+        ('a', (sym.IntLiteral(1),)): sym.IntLiteral(1),
+        ('a', (sym.IntLiteral(2),)): sym.IntLiteral(2),
+        ('a', (sym.IntLiteral(3),)): sym.IntLiteral(3),
+    }
+
+    analysis.get_attacher().visit(assignments[-1], constants_map=constants_map)
+
+    assert ('a', (sym.IntLiteral(1),)) not in constants_map
+    assert ('a', (sym.IntLiteral(2),)) not in constants_map
+    assert ('a', (sym.IntLiteral(3),)) not in constants_map
 
 
 def test_constant_propagation_transformer_export():
@@ -50,7 +135,7 @@ end subroutine const_prop_literals
     function()
 
     transformed = ConstantPropagationTransformer().visit(routine)
-    assignments = [str(a) for a in FindNodes(Assignment).visit(transformed.body)]
+    assignments = [str(a) for a in FindNodes(ir.Assignment).visit(transformed.body)]
 
     assert 'Assignment:: a = 1' in assignments
     assert 'Assignment:: b = 1.5' in assignments
@@ -94,7 +179,7 @@ end subroutine const_prop_ops_int
     assert outputs[4] == 0
 
     transformed = ConstantPropagationTransformer().visit(routine)
-    assignments = [str(a) for a in FindNodes(Assignment).visit(transformed.body)]
+    assignments = [str(a) for a in FindNodes(ir.Assignment).visit(transformed.body)]
 
     assert 'Assignment:: a_add = 3' in assignments
     assert 'Assignment:: a_sub = -1' in assignments
@@ -144,7 +229,7 @@ end subroutine test_constant_propagation_ops_bool_short_circuiting
     assert (outputs[1] == 1) is True
 
     transformed = ConstantPropagationTransformer().visit(routine)
-    assignments = [str(a) for a in FindNodes(Assignment).visit(transformed.body)]
+    assignments = [str(a) for a in FindNodes(ir.Assignment).visit(transformed.body)]
 
     assert 'Assignment:: a_and = False' in assignments
     assert 'Assignment:: a_or = True' in assignments
@@ -172,8 +257,8 @@ end subroutine test_constant_propagation_conditional_basic
 
     transformed = ConstantPropagationTransformer().visit(routine)
 
-    assert len(FindNodes(Conditional).visit(transformed.body)) == 1
-    assignments = [str(a) for a in FindNodes(Assignment).visit(transformed.body)]
+    assert len(FindNodes(ir.Conditional).visit(transformed.body)) == 1
+    assignments = [str(a) for a in FindNodes(ir.Assignment).visit(transformed.body)]
     assert 'Assignment:: c = 5' in assignments
     assert 'Assignment:: c = 3' in assignments
 
@@ -202,8 +287,8 @@ end subroutine test_constant_propagation_for_loop_basic
     # TODO: This should be internalised to auto-propagate on-demand
     transformed = ConstantPropagationTransformer().visit(transformed)
 
-    assert len(FindNodes(Loop).visit(transformed.body)) == 0
-    assignments = [str(a) for a in FindNodes(Assignment).visit(transformed.body)]
+    assert len(FindNodes(ir.Loop).visit(transformed.body)) == 0
+    assignments = [str(a) for a in FindNodes(ir.Assignment).visit(transformed.body)]
     for i in range(1, 6):
         assert f'Assignment:: c = {3*i}' in assignments
 
@@ -230,7 +315,7 @@ end subroutine test_constant_propagation_for_loop_basic_no_unroll
     routine = Subroutine.from_source(fcode, frontend=frontend)
     transformed = ConstantPropagationTransformer().visit(routine)
 
-    assignments = [str(a) for a in FindNodes(Assignment).visit(transformed.body)]
+    assignments = [str(a) for a in FindNodes(ir.Assignment).visit(transformed.body)]
     assert 'Assignment:: c = 15' in assignments
     assert 'Assignment:: d = 5*i' in assignments
     assert 'Assignment:: c = 30' in assignments
@@ -262,6 +347,6 @@ end subroutine test_constant_propagation_loop_nested_siblings_no_unroll
     routine = Subroutine.from_source(fcode, frontend=frontend)
     transformed = ConstantPropagationTransformer().visit(routine)
 
-    assignments = [str(a) for a in FindNodes(Assignment).visit(transformed.body)]
+    assignments = [str(a) for a in FindNodes(ir.Assignment).visit(transformed.body)]
     assert 'Assignment:: c = 5' in assignments
     assert 'Assignment:: c = 3' in assignments
